Refine Parser concept:
* Parser is not HTTP-specific
* parse algorithms are in their own header
* Update documentation
