+++ doc/types/FieldSequence.qbk
@@ -7,13 +7,13 @@
 
 [section:FieldSequence FieldSequence requirements]
 
-A [*`FieldSequence`] is an iterable container whose value type meets
-the requirements of [link beast.ref.Field [*`Field`]]. Objects meeting
-these requirements are serializable.
+A [*FieldSequence] is an iterable container whose value type meets
+the requirements of [link beast.ref.Field [*Field]]. Objects that meet
+these requirements become serializable by the implementation.
 
 In this table:
 
-* `X` denotes a type that meets the requirements of [*`FieldSequence`].
+* `X` denotes a type that meets the requirements of [*FieldSequence].
 
 * `a` is a value of type `X`.
 
@@ -23,14 +23,14 @@ In this table:
     [`X::value_type`]
     []
     [
-        A type that meets the requirements of `Field`.
+        A type that meets the requirements of @b Field.
     ]
 ]
 [
     [`X::const_iterator`]
     []
     [
-        A type that meets the requirements of `ForwardIterator`.
+        A type that meets the requirements of @b ForwardIterator.
     ]
 ]
 [
