+++ include/beast/http.hpp
@@ -15,6 +15,7 @@
 #include <beast/http/headers.hpp>
 #include <beast/http/message.hpp>
 #include <beast/http/message_v1.hpp>
+#include <beast/http/parse.hpp>
 #include <beast/http/parse_error.hpp>
 #include <beast/http/parser_v1.hpp>
 #include <beast/http/read.hpp>
